+++ src/collective/fingerpointing/upgrades/v4/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from collective.fingerpointing.config import PROJECTNAME
+from collective.fingerpointing.logger import logger
+
+
+def update_registry(setup_tool):
+    profile = 'profile-{0}:default'.format(PROJECTNAME)
+    setup_tool.runImportStepFromProfile(profile, 'registry')
+    logger.info('Registry updated.')